+++ types/src/bytesrepr.rs
@@ -17,6 +17,8 @@ use core::{
     ptr::NonNull,
 };
 
+#[cfg(feature = "datasize")]
+use datasize::DataSize;
 use num_integer::Integer;
 use num_rational::Ratio;
 use serde::{Deserialize, Serialize};
@@ -100,6 +102,7 @@ pub fn allocate_buffer<T: ToBytes>(to_be_serialized: &T) -> Result<Vec<u8>, Erro
 
 /// Serialization and deserialization errors.
 #[derive(Debug, PartialEq, Eq, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "datasize", derive(DataSize))]
 #[repr(u8)]
 pub enum Error {
     /// Early end of stream while deserializing.